fix: Sonatype group id should be `io.github.bric3`
Not `com.github.bric3`
